Put all conflicting plugins in a single class and clean up interfaces
